Refactor output content to OutputContent component
